Fix modal.
