+++ src/UniversalDashboard.UITest/TestFramework.psm1
@@ -3,7 +3,7 @@ Import-Module "$PSScriptRoot\..\Selenium\Selenium.psm1" -Force
 function Get-ModulePath {
     param([Switch]$Release)
 
-    $Global:Debug = (-not $Release.IsPresent)
+    $Env:Debug = (-not $Release.IsPresent)
     "$PSScriptRoot\..\output\UniversalDashboard.Community.psd1"
 }
